big update
- remove goal -> task
- refactor authentication bloc,...
- initial task feature, firebase messaging
- add notification
- fix bug

+++ lib/features/habit/data/data_sources/habit_data_source.dart
@@ -1,3 +1,4 @@
+import 'dart:developer' as dev;
 import 'dart:developer';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
@@ -7,35 +8,44 @@ import '../../../../core/constants/constants.dart';
 import '../../../../core/constants/typedef.dart';
 import '../../../../core/error/exceptions.dart';
 import '../../../../core/utils/helpers.dart';
-import '../../../authentication/data/data_sources/authentication_data_source.dart';
+import '../../../../core/notification/local_notification_manager.dart';
 import '../../../authentication/domain/entities/user_entity.dart';
 import '../../domain/entities/habit_entity.dart';
 import '../../domain/entities/streak_instance_entity.dart';
 import '../models/habit_instance_model.dart';
 import '../models/habit_model.dart';
 
+/// *deprecated: /users/ [user#email] / habits [habit#summary]
+// final habitCollectionRef = _firestore.collection(pathToUsers).doc(habit.creator!.email).collection(pathToHabits);
+
 abstract class HabitDataSource {
   SQuerySnapshot getHabitStream(UserEntity user);
+  SDocumentSnapshot getOneHabitStream(String hid);
   Future<void> addHabit(HabitEntity habit);
   Future<void> updateHabit(HabitEntity habit);
   Future<void> deleteHabit(HabitEntity habit);
 
   Future<HabitEntity?> getHabitByHid(String hid);
-  Future<List<StreakInstanceEntity>> getTopStreakOfHabit(HabitEntity habit);
+  Future<List<StreakInstanceEntity>> getTopHabitStreakOfUser(String hid, String collectionEmailPath);
+
+  Future<void> addMember(String hid, String email);
+  Future<void> removeMember(String hid, String email);
 }
 
 @Singleton(as: HabitDataSource)
 class HabitDataSourceImpl extends HabitDataSource {
-  HabitDataSourceImpl(this._firestore, this._authenticationDataSource);
+  HabitDataSourceImpl(this._firestore, this._localNotificationManager);
 
   final FirebaseFirestore _firestore;
-  final AuthenticationDataSource _authenticationDataSource;
+  final LocalNotificationManager _localNotificationManager;
 
   @override
   Future<void> addHabit(HabitEntity habit) async {
     try {
       // get the collection reference
-      final habitCollectionRef = _firestore.collection(pathToUsers).doc(habit.creator!.email).collection(pathToHabits);
+
+      // *new: /habits/ [habit#summary]
+      final habitCollectionRef = _firestore.collection(pathToHabits);
 
       // create a new document with a unique id
       final hid = habitCollectionRef.doc().id;
@@ -43,7 +53,17 @@ class HabitDataSourceImpl extends HabitDataSource {
       habit = habit.copyWith(hid: hid);
       // convert to the type that firestore accepts
       final habitDocument = (HabitModel.fromEntity(habit)).toDocument();
-      await habitCollectionRef.doc(hid).set(habitDocument);
+      await habitCollectionRef.doc(hid).set(habitDocument).then(
+        (value) {
+          // > Add notification for the habit
+          // ? check if habit use default reminder or not
+          if (habit.reminders!.useDefault) {
+            // set the notification same as the start
+            loadDailyNotification(habit, _localNotificationManager);
+          }
+          // < Add notification for the habit
+        },
+      );
     } on FirebaseException catch (e) {
       throw ServerException(code: e.code, message: e.toString());
     } catch (e) {
@@ -53,31 +73,60 @@ class HabitDataSourceImpl extends HabitDataSource {
 
   @override
   Future<void> deleteHabit(HabitEntity habit) async {
-    try {
-      final habitCollRef = _firestore.collection(pathToUsers).doc(habit.creator!.email).collection(pathToHabits);
+    // implement below just delete the document.
+    final habitCollRef = _firestore.collection(pathToHabits);
+    await habitCollRef.doc(habit.hid).delete();
 
-      await habitCollRef.doc(habit.hid).collection(pathToHabitInstances).get().then((snapshot) {
-        for (var ds in snapshot.docs) {
-          ds.reference.delete();
-        }
-      });
-      await habitCollRef.doc(habit.hid).delete();
-    } catch (e) {
-      log('rethrow --> Summary Exception: type: ${e.runtimeType.toString()} -- ${e.toString()}');
-      rethrow;
+    // > cancel the notification
+    _localNotificationManager.I.cancel(habit.reminders!.rid!);
+
+    // > delete all sub collection
+
+    // <https://firebase.google.com/docs/firestore/manage-data/delete-data>
+    // loop for all sub collection has id = email
+    if (habit.members!.isNotEmpty) {
+      for (var member in habit.members!) {
+        final habitInstanceCollRef = habitCollRef.doc(habit.hid).collection(member);
+        await habitInstanceCollRef.get().then((snapshot) {
+          for (var ds in snapshot.docs) {
+            ds.reference.delete();
+          }
+        });
+      }
     }
+    // < delete all sub collection
+
+    // await habitCollRef.doc(habit.hid).delete();
   }
 
   @override
   Future<void> updateHabit(HabitEntity habit) async {
     try {
-      final habitCollectionRef = _firestore.collection(pathToUsers).doc(habit.creator!.email).collection(pathToHabits);
+      // final habitCollectionRef = _firestore.collection(pathToUsers).doc(habit.creator!.email).collection(pathToHabits);
+      final habitCollectionRef = _firestore.collection(pathToHabits);
 
       final habitModel = HabitModel.fromEntity(habit);
 
-      await habitCollectionRef.doc(habit.hid).update(habitModel.toDocument());
+      await habitCollectionRef.doc(habit.hid).update(habitModel.toDocument()).then((value) {
+        // > Edit notification for the habit
+        // ? check if habit use default reminder or not
+        if (habit.reminders!.useDefault) {
+          // set the notification same as the start
+          loadDailyNotification(habit, _localNotificationManager);
+        }
+        // < Edit notification for the habit
+      });
+
+      // update for each habit instance that has edited filed is false (mean that instance name like parent habit)
+      final habitInstanceCollectionRef =
+          _firestore.collection(pathToHabits).doc(habit.hid).collection(habit.creator!.email!).where('edited', isEqualTo: false);
+      await habitInstanceCollectionRef.get().then((snapshot) {
+        for (var ds in snapshot.docs) {
+          ds.reference.update({'summary': habit.summary});
+        }
+      });
     } on FirebaseException catch (e) {
-      log('rethrow --> Summary Exception: type: ${e.runtimeType.toString()} -- ${e.toString()}');
+      dev.log('rethrow --> Summary Exception: type: ${e.runtimeType.toString()} -- ${e.toString()}');
       rethrow;
     } catch (e) {
       throw ServerException(message: e.toString());
@@ -86,38 +135,63 @@ class HabitDataSourceImpl extends HabitDataSource {
 
   @override
   SQuerySnapshot getHabitStream(UserEntity user) {
-    final habitCollRef = _firestore.collection(pathToUsers).doc(user.email).collection(pathToHabits).orderBy('summary');
+    // final habitCollRef = _firestore.collection(pathToUsers).doc(user.email).collection(pathToHabits).orderBy('summary');
+    dev.log('getHabitStream params user: $user');
+    // get all habits that user is owner & member
+    final habitCollRef = _firestore
+        .collection(pathToHabits)
+        // .where('creator.email', isEqualTo: user.email)
+        .where(Filter.or(
+          Filter('creator.email', isEqualTo: user.email),
+          // Filter('members', arrayContains: UserModel.fromEntity(user).toDocument()),
+          Filter('members', arrayContains: user.email),
+        ))
+        .orderBy('summary');
+
+    // > retrieve notification for each habit to set the local notification
+    habitCollRef.get().then((snapshot) {
+      if (snapshot.docs.isEmpty) {
+        _localNotificationManager.I.cancelAll();
+        return habitCollRef.snapshots();
+      }
+      
+      for (var doc in snapshot.docs) {
+        final HabitModel habit = HabitModel.fromDocument(doc.data());
+        // ? check if habit use default reminder or not
+        if (habit.reminders!.useDefault) {
+          // set the notification same as the start
+          loadDailyNotification(habit, _localNotificationManager);
+        }
+      }
+    });
     return habitCollRef.snapshots();
   }
 
   @override
   Future<HabitEntity?> getHabitByHid(String hid) async {
-    final user = _authenticationDataSource.currentUser;
-    if (user != null) {
-      final habitDocRef = _firestore.collection(pathToUsers).doc(user.email).collection(pathToHabits).doc(hid);
-      final habitDoc = await habitDocRef.get();
+    // final habitDocRef = _firestore.collection(pathToUsers).doc(user.email).collection(pathToHabits).doc(hid);
+    final habitDocRef = _firestore.collection(pathToHabits).doc(hid);
+    final habitDoc = await habitDocRef.get();
 
-      if (habitDoc.exists) {
-        return HabitModel.fromDocument(habitDoc.data()!);
-      } else {
-        throw ServerException(message: 'Habit not found');
-      }
+    if (habitDoc.exists) {
+      return HabitModel.fromDocument(habitDoc.data()!);
+    } else {
+      // throw ServerException(message: 'Habit not found');
+      return null;
     }
-    return null;
   }
 
   // scenarios:
   // 1. the first day of the list is completed [1,0...]
   // 2. the middle day of the list is completed [...0,1,0...]
   // 3. the last day of the list is completed [...0,1]
+
+  //? the main params is path to habit instances collection [hid, email] -- /habits/ [hid] / [email]
   @override
-  Future<List<StreakInstanceEntity>> getTopStreakOfHabit(HabitEntity habit) async {
-    final habitICollRef = _firestore
-        .collection(pathToUsers)
-        .doc(habit.creator!.email)
-        .collection(pathToHabits)
-        .doc(habit.hid)
-        .collection(pathToHabitInstances);
+  Future<List<StreakInstanceEntity>> getTopHabitStreakOfUser(String hid, String collectionEmailPath) async {
+    // final habitICollRef =
+    //     _firestore.collection(pathToUsers).doc(habit.creator!.email).collection(pathToHabits).doc(habit.hid).collection(pathToHabitInstances);
+    final habitICollRef = _firestore.collection(pathToHabits).doc(hid).collection(collectionEmailPath);
     final habitIDocRef = await habitICollRef.orderBy('date').get();
 
     List<HabitInstanceModel> instances = [];
@@ -128,25 +202,27 @@ class HabitDataSourceImpl extends HabitDataSource {
       return [];
     }
 
-    List<StreakInstanceEntity> result = [];
-    int currentStreakLength = 0;
-    DateTime currentStreakStart = instances.first.date!;
+    List<StreakInstanceEntity> result = []; // each StreakInstanceEntity contain [start, end, length]
+    int currentStreakLength = 0; // the length of current streak in loop
+    DateTime currentStreakStart = instances.first.date!; // the start date of current streak in loop
+    //? cannot move [currentStreakStart] into the loop because it will be override by the next day
 
+    // loop through all instances to find the top streaks
     for (int i = 0; i < instances.length; i++) {
-      final HabitInstanceModel instance = instances[i];
+      final HabitInstanceModel instance = instances[i]; // current instance
       final DateTime currentStreakDate = instance.date!;
-      final bool isLastInstance = (i == instances.length - 1);
-      final bool isStreakCompleted = instance.completed!;
+      final bool isLastInstance = (i == instances.length - 1); // check if the current instance is the last one <-- length - 1
+      final bool isInstanceCompleted = instance.completed!; // current instance is completed or not
 
-      if (isStreakCompleted) {
+      if (isInstanceCompleted) {
         currentStreakLength++;
-        // check if the next day has init instance or not
+        // the next day has init instance or not
         final isCreated = await isCreatedInstance(
-          iid: getIid(habit.hid!, currentStreakDate.add(const Duration(days: 1))),
+          iid: getIid(hid, currentStreakDate.add(const Duration(days: 1))),
           collectionReference: habitICollRef,
         );
 
-        // if the next day doesn't have init instance, then...
+        // if the next day doesn't have init instance, means that the streak is broken -> add the streak to result
         if (!isCreated) {
           final StreakInstanceEntity streak = StreakInstanceEntity(
             start: currentStreakStart,
@@ -158,7 +234,10 @@ class HabitDataSourceImpl extends HabitDataSource {
           if (!isLastInstance) currentStreakStart = instances[i + 1].date!;
         }
       } else {
-        // the streak is broken
+        // the current instance is not completed
+
+        // if the current streak length > 0, means that before this instance, there is some completed instances
+        // -> add the streak to result
         if (currentStreakLength > 0) {
           final StreakInstanceEntity streak = StreakInstanceEntity(
             start: currentStreakStart,
@@ -170,6 +249,9 @@ class HabitDataSourceImpl extends HabitDataSource {
         }
       }
 
+      // if the current instance is the last one and the current streak length > 0
+      // -- the last streak is not added to result yet
+      // -> add the streak to result
       if (isLastInstance && currentStreakLength > 0) {
         final StreakInstanceEntity streak = StreakInstanceEntity(
           start: currentStreakStart,
@@ -179,16 +261,53 @@ class HabitDataSourceImpl extends HabitDataSource {
         result.add(streak);
       }
 
-      if (!isStreakCompleted && !isLastInstance) {
-        currentStreakStart = instances[i + 1].date!;
+      // if the current instance is not completed && not the last one
+      if (!isInstanceCompleted && !isLastInstance) {
+        currentStreakStart = instances[i + 1].date!; // move to the next day
       }
     }
+
+    // sort the result by length
     result.sort((prev, next) => next.length.compareTo(prev.length));
     if (result.length > 3) {
       result = result.sublist(0, 3);
     }
     result.reversed;
     return result;
+  }
+
+  @override
+  SDocumentSnapshot getOneHabitStream(String hid) {
+    final habitCollRef = _firestore
+        .collection(pathToHabits)
+        // .where('creator.email', isEqualTo: user.email)
+        .doc(hid);
+    return habitCollRef.snapshots();
+  }
+
+  @override
+  Future<void> addMember(String hid, String email) async {
+    try {
+      _firestore.collection(pathToHabits).doc(hid).update({
+        'members': FieldValue.arrayUnion([email])
+      }).then((value) => log('Add Member Success'));
+    } catch (e) {
+      log('Summary Exception: type: ${e.runtimeType.toString()} -- ${e.toString()}');
+    }
+  }
+
+  @override
+  Future<void> removeMember(String hid, String email) async {
+    try {
+      await _firestore.collection(pathToHabits).doc(hid).update({
+        'members': FieldValue.arrayRemove([email])
+      }).then((value) => log('Remove Member Success'));
+      // NOTE >? should delete all habit instances of the member ?
+
+      return;
+    } catch (e) {
+      log('Summary Exception: type: ${e.runtimeType.toString()} -- ${e.toString()}');
+    }
   }
 }
 
@@ -197,4 +316,18 @@ Future<bool> isCreatedInstance({
   required CollectionReference<Map<String, dynamic>> collectionReference,
 }) async {
   return await collectionReference.doc(iid).get().then((value) => value.exists);
+}
+
+Future<void> loadDailyNotification(HabitEntity habit, LocalNotificationManager localNotificationManager) async {
+  // ? check if habit use default reminder or not
+  if (habit.reminders!.useDefault) {
+    // set the notification same as the start
+    localNotificationManager.setDailyScheduleNotification(
+      id: habit.reminders!.rid!,
+      title: 'Time to ${habit.summary}',
+      body: habit.description,
+      scheduledDate: habit.start!,
+      payload: habit.hid,
+    );
+  }
 }
